lambda-functions/image-processor/requirements.txt
Pillow==12.0.0
boto3==1.34.0
